refactor: :recycle: refactor some code
